refactor : 가격을 int 타입에서 BigDecimal 타입으로 변경

+++ goatgam/src/main/java/com/sparta/goatgam/domain/owner/dto/FoodOptionRequestDto.java
@@ -3,11 +3,11 @@ package com.sparta.goatgam.domain.owner.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.util.UUID;
+import java.math.BigDecimal;
 
 @Getter
 @AllArgsConstructor
 public class FoodOptionRequestDto {
     private String contents;
-    private int surcharge;
+    private BigDecimal surcharge;
 }
